capicxx-core-tools: Update to 3.1.12.3
N.B. These recipes are set up to download binary releases.

It should be updated to 3.1.12.3 because 3.1.12 (the .0 is omitted
apparently) is an old version.  A rename is sufficient because the
${PV} controls which release is downloaded.

Dropped dual checksums - it's unnecessary and just adds maintenance effort

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>
(cherry picked from commit 61180da7880e66c9924599e135dd8f896255f3a3)